requirements.txt
Flask==1.0.2
requests==2.18.4
psycopg2-binary==2.7.5
Flask_SQLAlchemy==2.3.2
jac==0.17.1
beautifulsoup4==4.6.0
django-htmlmin==0.10.0
gunicorn==19.8.1
flask-compress==1.4.0
